deps/v8/test/webkit/fast/js/regexp-bol-expected.txt

# Copyright 2013 the V8 project authors. All rights reserved.
# Copyright (C) 2005, 2006, 2007, 2008, 2009 Apple Inc. All rights reserved.
#
# Redistribution and use in source and binary forms, with or without
# modification, are permitted provided that the following conditions
# are met:
# 1.  Redistributions of source code must retain the above copyright
#     notice, this list of conditions and the following disclaimer.
# 2.  Redistributions in binary form must reproduce the above copyright
#     notice, this list of conditions and the following disclaimer in the
#     documentation and/or other materials provided with the distribution.
#
# THIS SOFTWARE IS PROVIDED BY APPLE INC. AND ITS CONTRIBUTORS ``AS IS'' AND ANY
# EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
# WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
# DISCLAIMED. IN NO EVENT SHALL APPLE INC. OR ITS CONTRIBUTORS BE LIABLE FOR ANY
# DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES
# (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
# LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON
# ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
# (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
# SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.

Test for beginning of line (BOL or ^) matching

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS s.match(/^notHere/) is null
PASS s.match(/^abc/) is ["abc"]
PASS s.match(/(^|X)abc/) is ["abc",""]
PASS s.match(/^longer|123/) is ["123"]
PASS s.match(/(^abc|c)123/) is ["abc123","abc"]
PASS s.match(/(c|^abc)123/) is ["abc123","abc"]
PASS s.match(/(^ab|abc)123/) is ["abc123","abc"]
PASS s.match(/(bc|^abc)([0-9]*)a/) is ["bc789a","bc","789"]
PASS /(?:(Y)X)|(X)/.exec("abc") is null
PASS /(?:(?:^|Y)X)|(X)/.exec("abc") is null
PASS /(?:(?:^|Y)X)|(X)/.exec("abcd") is null
PASS /(?:(?:^|Y)X)|(X)/.exec("Xabcd") is ["X",undefined]
PASS /(?:(?:^|Y)X)|(X)/.exec("aXbcd") is ["X","X"]
PASS /(?:(?:^|Y)X)|(X)/.exec("abXcd") is ["X","X"]
PASS /(?:(?:^|Y)X)|(X)/.exec("abcXd") is ["X","X"]
PASS /(?:(?:^|Y)X)|(X)/.exec("abcdX") is ["X","X"]
PASS /(?:(?:^|Y)X)|(X)/.exec("YXabcd") is ["YX",undefined]
PASS /(?:(?:^|Y)X)|(X)/.exec("aYXbcd") is ["YX",undefined]
PASS /(?:(?:^|Y)X)|(X)/.exec("abYXcd") is ["YX",undefined]
PASS /(?:(?:^|Y)X)|(X)/.exec("abcYXd") is ["YX",undefined]
PASS /(?:(?:^|Y)X)|(X)/.exec("abcdYX") is ["YX",undefined]
PASS successfullyParsed is true

TEST COMPLETE
